LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/muted-errors.sub-expected.txt

CONSOLE MESSAGE: fox
CONSOLE MESSAGE: fox
CONSOLE MESSAGE: fox
CONSOLE MESSAGE: fox
CONSOLE MESSAGE: fox
CONSOLE MESSAGE: fox
CONSOLE MESSAGE: SyntaxError: ':::not-going-to-be-valid' is not a valid selector.
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'cannot'


PASS Errors for same-origin script shouldn't be muted
PASS Errors for cross-origin script should be muted
PASS Errors for cross-origin script should be muted even if the script is once loaded as same-origin
PASS Errors for same-origin scripts redirected to a cross-origin url and redirected back to same-origin should be muted
PASS Errors for cross-origin scripts redirected to a same-origin url should be muted
PASS Non-synthetic errors for same-origin scripts redirected to a cross-origin URL and redirected back to same-origin should be muted
PASS Syntax error for same-origin script redirected to a cross-origin URL and redirected back to same-origin should be muted
